基本クラスを継承するようにした。
