+++ Assets/Objects/Player/PlayerAction/DriftPlayerActionBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6943ffd66ad50aa42b380fb188176893
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
